requirements.txt
Package      Version
------------ -------------------
brotlipy     0.7.0
certifi      2020.12.5
cffi         1.14.5
chardet      4.0.0
click        7.1.2
cryptography 3.3.1
Flask        1.1.2
Flask-Cors   3.0.10
idna         2.10
itsdangerous 1.1.0
Jinja2       2.11.3
MarkupSafe   1.1.1
pip          21.0.1
pycparser    2.20
pycrypto     2.6.1
pyOpenSSL    20.0.1
PySocks      1.7.1
requests     2.25.1
setuptools   52.0.0.post20210125
six          1.15.0
urllib3      1.26.3
Werkzeug     1.0.1
wheel        0.36.2
